+++ src/components/TouchBar.js
@@ -5,7 +5,7 @@ import { useEffect } from 'react';
 import {
 	TOUCH_BAR_FORMAT_BUTTON_TOUCHED,
 	TOUCH_BAR_SELECT_SERVER_TOUCHED,
-} from './actions';
+} from '../scripts/actions';
 
 const useSelectServerPanel = (activeServerUrl, servers, dispatch) => {
 	class SelectServerPanel {
